requirements.txt
Django==1.7.6
djangorestframework==3.1.0
Pillow==2.7.0
psycopg2==2.6
requests==2.6.0
